requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
Flask-Mail==0.9.1
python-dotenv==1.0.0
geopy==2.3.0
